+++ lib/Castle/Authenticate.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Castle;
+
+class Authenticate extends RestModel
+{
+  protected $isSingular = true;
+}